refactored class name and CardManagementPageInitialize extended IyzipayResource instead of CardManagementPageResource.
